+++ client/flutter/fadzmaq/lib/controllers/request.dart
@@ -1,13 +1,8 @@
 import 'dart:async';
-import 'package:fadzmaq/controllers/cache.dart';
 import 'package:fadzmaq/models/app_config.dart';
-import 'package:fadzmaq/models/matches.dart';
-import 'package:fadzmaq/models/matches.dart' as prefix0;
-import 'package:fadzmaq/models/recommendations.dart';
 import 'package:fadzmaq/views/loginscreen.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:google_sign_in/google_sign_in.dart';
 import 'package:http/http.dart' as http;
 import 'dart:convert';
 import 'package:fadzmaq/models/models.dart';
