+++ client/flutter/fadzmaq/lib/views/matches.dart
@@ -1,11 +1,9 @@
 import 'package:fadzmaq/controllers/cache.dart';
-import 'package:fadzmaq/controllers/globals.dart';
+import 'package:fadzmaq/controllers/globalData.dart';
 import 'package:fadzmaq/models/mainModel.dart';
-import 'package:fadzmaq/models/profile.dart';
 import 'package:fadzmaq/views/widgets/matchEntry.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/foundation.dart';
-import 'package:fadzmaq/controllers/request.dart';
 import 'package:fadzmaq/models/matches.dart';
 
 class MatchesPage extends StatelessWidget {
@@ -13,7 +11,7 @@ class MatchesPage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return LoadModel(
+    return VerifyModel(
         model: Model.matches,
         builder: (context) {
           return MatchesList();
@@ -41,7 +39,7 @@ class MatchesList extends StatelessWidget {
 
   Widget _listItemBuilder(BuildContext context, int index) {
     // MatchesData matchesData = RequestProvider.of<MatchesData>(context);
-    MatchesData matchesData = DataController.of(context).matches;
+    MatchesData matchesData = GlobalData.of(context).matches;
     return MatchEntry(profile: matchesData.matches[index].profile);
   }
 }
