+++ client/flutter/fadzmaq/lib/views/widgets/deleteUser.dart
@@ -2,9 +2,7 @@ import 'package:fadzmaq/controllers/postAsync.dart';
 import 'package:fadzmaq/views/loginscreen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:flutter/foundation.dart';
 import 'package:google_sign_in/google_sign_in.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:http/http.dart' as http;
 
 void deleteUser(BuildContext context) async {
