+++ client/flutter/fadzmaq/lib/views/widgets/displayPhoto.dart
@@ -1,5 +1,3 @@
-import 'dart:ui';
-
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:fadzmaq/controllers/cache.dart';
 import 'package:flutter/material.dart';
