+++ client/flutter/fadzmaq/lib/views/widgets/hobbyChips.dart
@@ -1,6 +1,5 @@
-import 'package:fadzmaq/controllers/request.dart';
 import 'package:fadzmaq/models/hobbies.dart';
-import 'package:fadzmaq/models/mainModel.dart';
+import 'package:fadzmaq/controllers/globalData.dart';
 import 'package:fadzmaq/models/profile.dart';
 import 'package:fadzmaq/views/widgets/hobbyChip.dart';
 import 'package:flutter/material.dart';
@@ -35,7 +34,7 @@ class HobbyChips extends StatelessWidget {
 
     // any request using hobby chips should have a user profile container request
     // to compare the hobbies to.
-    ProfileData userProfile = DataController.of(context).userProfile;
+    ProfileData userProfile = GlobalData.of(context).userProfile;
 
     if (userProfile == null) return Container();
     if (userProfile.hobbyContainers == null) return Container();
