+++ client/flutter/fadzmaq/lib/views/widgets/matchDialog.dart
@@ -1,5 +1,5 @@
 import 'package:fadzmaq/controllers/globals.dart';
-import 'package:fadzmaq/models/mainModel.dart';
+import 'package:fadzmaq/controllers/globalData.dart';
 import 'package:fadzmaq/models/profile.dart';
 import 'package:fadzmaq/views/landing.dart';
 import 'package:fadzmaq/views/profilepage.dart';
@@ -8,7 +8,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
 Future<void> matchPopup(BuildContext context, ProfileData profile) async {
-  ProfileData userProfile = DataController.of(context).userProfile;
+  ProfileData userProfile = GlobalData.of(context).userProfile;
 
   return showDialog<void>(
     context: mainScaffold.currentContext,
