+++ client/flutter/fadzmaq/lib/views/widgets/recommendationEntry.dart
@@ -1,6 +1,5 @@
 import 'package:fadzmaq/controllers/globals.dart';
 import 'package:fadzmaq/controllers/postAsync.dart';
-import 'package:fadzmaq/controllers/request.dart';
 import 'package:fadzmaq/models/profile.dart';
 import 'package:fadzmaq/views/profilepage.dart';
 import 'package:fadzmaq/views/recommendations.dart';
